% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{fizz_buzz}
\alias{fizz_buzz}
\title{Create n items of the "FizzBuzz" series.}
\usage{
fizz_buzz(n)
}
\arguments{
\item{n}{A non-zero positive integer representing the number of items to return}
}
\value{
A vector of integers and strings following the rules of the
"FizzBuzz" series
}
\description{
The FizzBuzz series replaces all multiples of 3 with "Fizz"
and all multiples of 5 with "Buzz", and multiples of both with
"Fizz-Buzz". All other integers remain as-is in a list.
}
\examples{
fizz_buzz(5)
fizz_buzz(22)
}
