% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A series of weather observations from Pulliam Airport Weather Station
in Flagstaff, AZ. Includes precipitation, snow, and high and low temps.}
\format{
a data frame with 25449 observations
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{PRCP}{Daily precipitation in inches.}
  \item{SNOW}{Daily snowfall amount in inches.}
  \item{TMAX}{Daily maximum temperature in degrees Fahrenheit.}
  \item{TMIN}{Daily minimum temperature in degrees Fahrenheit.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A series of weather observations from Pulliam Airport Weather Station
in Flagstaff, AZ. Includes precipitation, snow, and high and low temps.
}
\keyword{datasets}
